pyproject.toml
[project]
name = "deadline-bot"
version = "0.1.0"
description = "Discord bot for research group deadline management"
requires-python = ">=3.11"
dependencies = [
    "discord.py>=2.4",
    "sqlmodel>=0.0.21",
    "aiosqlite>=0.20",
    "apscheduler>=3.10",
    "httpx>=0.27",
    "python-dotenv>=1.0",
    "python-dateutil>=2.9",
    "pydantic-settings>=2.3",
    "greenlet>=3.0",
    "alembic>=1.13",
    "fastapi>=0.111",
    "uvicorn[standard]>=0.30",
]

[dependency-groups]
dev = [
    "ruff>=0.4",
    "ty>=0.0.1a1",
    "pytest>=8.2",
    "pytest-asyncio>=0.23",
    "pytest-mock>=3.14",
    "pytest-cov>=5.0",
    "respx>=0.21",
]

[tool.ruff]
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.coverage.run]
source = ["."]
omit = ["tests/*", ".venv/*"]
